Separando correctamente protocolo con server/client
